+++ packages/features/src/Domain/Component/ComponentFlag.ts
@@ -1,0 +1,4 @@
+export enum ComponentFlag {
+  New = 'New',
+  Deprecated = 'Deprecated',
+}