+++ packages/features/src/Domain/Component/ComponentPermission.ts
@@ -1,0 +1,7 @@
+import { ContentType } from '@standardnotes/common'
+import { ComponentAction } from './ComponentAction'
+
+export type ComponentPermission = {
+  name: ComponentAction
+  content_types?: ContentType[]
+}