+++ packages/features/src/Domain/Component/ThemeDockIcon.ts
@@ -1,0 +1,7 @@
+export type ThemeDockIcon = {
+  type: 'svg' | 'circle'
+  background_color: string
+  foreground_color: string
+  border_color: string
+  source?: string
+}