+++ packages/features/src/Domain/index.ts
@@ -1,0 +1,13 @@
+export * from './Feature/FeatureDescription'
+export * from './Feature/FeatureIdentifier'
+export * from './Feature/Features'
+
+export * from './Permission/Permission'
+export * from './Permission/PermissionName'
+
+export * from './Component/ComponentAction'
+export * from './Component/ComponentArea'
+export * from './Component/ComponentFlag'
+export * from './Component/ComponentPermission'
+export * from './Component/NoteType'
+export * from './Component/ThemeDockIcon'