+++ packages/encryption/CHANGELOG.md
@@ -3,6 +3,16 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.10.6](https://github.com/standardnotes/app/compare/@standardnotes/encryption@1.10.5...@standardnotes/encryption@1.10.6) (2022-07-05)
+
+**Note:** Version bump only for package @standardnotes/encryption
+
+## [1.10.5](https://github.com/standardnotes/app/compare/@standardnotes/encryption@1.10.4...@standardnotes/encryption@1.10.5) (2022-07-05)
+
+### Bug Fixes
+
+* add typescript dev dependency on encryption package ([175a0ff](https://github.com/standardnotes/app/commit/175a0ff7db7a83e67bf1991dd46eff1a7c7920ff))
+
 ## [1.10.4](https://github.com/standardnotes/app/compare/@standardnotes/encryption@1.10.3...@standardnotes/encryption@1.10.4) (2022-07-05)
 
 ### Bug Fixes