+++ packages/web-server/CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.2.63](https://github.com/standardnotes/app/compare/@standardnotes/web-server@1.2.62...@standardnotes/web-server@1.2.63) (2022-07-05)
+
+**Note:** Version bump only for package @standardnotes/web-server
+
+## [1.2.62](https://github.com/standardnotes/app/compare/@standardnotes/web-server@1.2.61...@standardnotes/web-server@1.2.62) (2022-07-05)
+
+**Note:** Version bump only for package @standardnotes/web-server
+
+## [1.2.61](https://github.com/standardnotes/app/compare/@standardnotes/web-server@1.2.60...@standardnotes/web-server@1.2.61) (2022-07-05)
+
+**Note:** Version bump only for package @standardnotes/web-server
+
+## [1.2.60](https://github.com/standardnotes/app/compare/@standardnotes/web-server@1.2.59...@standardnotes/web-server@1.2.60) (2022-07-05)
+
+**Note:** Version bump only for package @standardnotes/web-server
+
 ## [1.2.59](https://github.com/standardnotes/app/compare/@standardnotes/web-server@1.2.58...@standardnotes/web-server@1.2.59) (2022-07-05)
 
 **Note:** Version bump only for package @standardnotes/web-server