+++ packages/web/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [3.33.6](https://github.com/standardnotes/app/compare/@standardnotes/web@3.33.5...@standardnotes/web@3.33.6) (2022-07-05)
+
+**Note:** Version bump only for package @standardnotes/web
+
+## [3.33.5](https://github.com/standardnotes/app/compare/@standardnotes/web@3.33.4...@standardnotes/web@3.33.5) (2022-07-05)
+
+**Note:** Version bump only for package @standardnotes/web
+
 ## [3.33.4](https://github.com/standardnotes/app/compare/@standardnotes/web@3.33.3...@standardnotes/web@3.33.4) (2022-07-05)
 
 **Note:** Version bump only for package @standardnotes/web